+++ GHEtool/gui/data_2_borefield_func.py
@@ -12,7 +12,7 @@ from GHEtool import Borefield, FluidData, GroundData, PipeData
 from GHEtool.gui.gui_structure import load_data_GUI
 
 if TYPE_CHECKING:  # pragma: no cover
-    from GHEtool.gui.gui_data_storage import DataStorage
+    from GHEtool.gui.gui_classes.gui_data_storage import DataStorage
     from numpy.typing import NDArray
